google_adk_voice_agents/pyproject.toml
[project]
name = "google-adk-voice-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.17.0",
    "yfinance>=0.2.66",
]
